Move splitgraph.yml-specific models into a separate module
